# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AstroHealth API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cultivations:
    get:
      operationId: listCultivations
      tags: [cultivations]
      summary: List cultivations with today's check-in embedded
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, paused, completed]
          required: false
      responses:
        "200":
          description: List of cultivations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CultivationWithTodayCheckIn"
    post:
      operationId: createCultivation
      tags: [cultivations]
      summary: Create a new cultivation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCultivationBody"
      responses:
        "201":
          description: Created cultivation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cultivation"
  /cultivations/{id}:
    patch:
      operationId: updateCultivation
      tags: [cultivations]
      summary: Update a cultivation
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCultivationBody"
      responses:
        "200":
          description: Updated cultivation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cultivation"
        "404":
          description: Not found
    delete:
      operationId: deleteCultivation
      tags: [cultivations]
      summary: Delete a cultivation (and its check-ins)
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /cultivations/{id}/check-ins/today:
    get:
      operationId: getCultivationTodayCheckIn
      tags: [cultivations]
      summary: Get today's check-in for a cultivation
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Today's check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CultivationCheckIn"
        "404":
          description: No check-in for today
  /cultivations/{id}/check-ins:
    get:
      operationId: listCultivationCheckIns
      tags: [cultivations]
      summary: List check-in history for a cultivation
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 30
          required: false
          description: "Number of past days to return (max 90, default 30)"
      responses:
        "200":
          description: List of check-ins ordered by date descending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CultivationCheckIn"
        "404":
          description: Cultivation not found
    post:
      operationId: upsertCultivationCheckIn
      tags: [cultivations]
      summary: Create or update a check-in for a cultivation
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCultivationCheckInBody"
      responses:
        "200":
          description: Upserted check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CultivationCheckIn"
  /cultivations/element-balance:
    get:
      operationId: getCultivationElementBalance
      tags: [cultivations]
      summary: Elemental balance of active cultivations over a date range
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            enum: [7, 30]
            default: 7
          required: false
          description: "Time window in days (7 or 30)"
      responses:
        "200":
          description: Element tally for active cultivations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElementBalance"
components:
  schemas:
...
    Cultivation:
      type: object
      properties:
        id:
          type: integer
        testerId:
          type: string
        title:
          type: string
        domain:
          type: string
        element:
          type: ["string", "null"]
          enum: ["fire", "earth", "air", "water", null]
          description: "Classical element derived from domain, or user override. Null if unassigned."
        description:
          type: ["string", "null"]
        relatedPlanet:
          type: ["string", "null"]
        relatedHouse:
          type: ["integer", "null"]
        relatedBodyWeatherTags:
          type: ["array", "null"]
          items:
            type: string
        targetPractice:
          type: ["string", "null"]
        frequency:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, testerId, title, domain, frequency, status, createdAt, updatedAt]
    CultivationCheckIn:
      type: object
      properties:
        id:
          type: integer
        testerId:
          type: string
        cultivationId:
          type: integer
        date:
          type: string
        completed:
          type: boolean
        effortLevel:
          type: ["integer", "null"]
          minimum: 1
          maximum: 5
        note:
          type: ["string", "null"]
        practicesCompleted:
          type: ["array", "null"]
          items:
            type: string
          description: "Named sub-practices completed in this session."
        durationMinutes:
          type: ["integer", "null"]
          description: "Optional total session duration in minutes."
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, testerId, cultivationId, date, completed, createdAt, updatedAt]
    CultivationWithTodayCheckIn:
      allOf:
        - $ref: "#/components/schemas/Cultivation"
        - type: object
          properties:
            todayCheckIn:
              oneOf:
                - $ref: "#/components/schemas/CultivationCheckIn"
                - type: "null"
    CreateCultivationBody:
      type: object
      properties:
        title:
          type: string
        domain:
          type: string
        element:
          type: ["string", "null"]
          enum: ["fire", "earth", "air", "water", null]
          description: "Override the auto-derived element. Omit to auto-derive from domain."
        description:
          type: ["string", "null"]
        relatedPlanet:
          type: ["string", "null"]
        relatedHouse:
          type: ["integer", "null"]
        relatedBodyWeatherTags:
          type: ["array", "null"]
          items:
            type: string
        targetPractice:
          type: ["string", "null"]
        frequency:
          type: string
          default: daily
        status:
          type: string
          enum: [active, paused, completed]
          default: active
      required: [title, domain]
    UpdateCultivationBody:
      type: object
      properties:
        title:
          type: string
        domain:
          type: string
        element:
          type: ["string", "null"]
          enum: ["fire", "earth", "air", "water", null]
          description: "Override the element. Pass null to clear. Omit to leave unchanged."
        description:
          type: ["string", "null"]
        relatedPlanet:
          type: ["string", "null"]
        relatedHouse:
          type: ["integer", "null"]
        relatedBodyWeatherTags:
          type: ["array", "null"]
          items:
            type: string
        targetPractice:
          type: ["string", "null"]
        frequency:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
    UpsertCultivationCheckInBody:
      type: object
      properties:
        completed:
          type: boolean
        effortLevel:
          type: ["integer", "null"]
          minimum: 1
          maximum: 5
        note:
          type: ["string", "null"]
        date:
          type: string
          description: "YYYY-MM-DD. Defaults to today. Past 30 days only; future dates rejected."
        practicesCompleted:
          type: ["array", "null"]
          items:
            type: string
          description: "Named sub-practices done this session. Non-empty array implies completed=true unless overridden."
        durationMinutes:
          type: ["integer", "null"]
          description: "Optional total session duration in minutes."
      required: []
    ElementBalanceEntry:
      type: object
      properties:
        tended:
          type: integer
          description: "Number of completed check-ins for this element in the window."
        total:
          type: integer
          description: "Maximum possible tending slots (active cultivations × days)."
      required: [tended, total]
    ElementBalance:
      type: object
      properties:
        fire:
          $ref: "#/components/schemas/ElementBalanceEntry"
        earth:
          $ref: "#/components/schemas/ElementBalanceEntry"
        air:
          $ref: "#/components/schemas/ElementBalanceEntry"
        water:
          $ref: "#/components/schemas/ElementBalanceEntry"
        unassigned:
          $ref: "#/components/schemas/ElementBalanceEntry"
      required: [fire, earth, air, water, unassigned]

    # ── Body Weather ─────────────────────────────────────────────────────────────
...